Move Apply controllers into their own namespace
